pyproject.toml
[tool.poetry]
name = "wodmaker"
version = "0.1.0"
description = ""
authors = ["Ivan Zugnoni <zugnoni.ivan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.5"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
